+++ src/usdb_syncer/gui/mw.py
@@ -149,6 +149,7 @@ class MainWindow(Ui_MainWindow, QMainWindow):
             (self.action_show_log, lambda: open_file_explorer(AppPaths.log)),
             (self.action_show_in_usdb, self._show_current_song_in_usdb),
             (self.action_open_song_folder, self._open_current_song_folder),
+            (self.action_delete, self.table.delete_selected_songs),
         ):
             action.triggered.connect(func)