+++ CHANGELOG.md
@@ -7,6 +7,8 @@
 ## Features
 
 - The batch view has been removed, and a view dedicated to local songs added.
+- Added several actions for downloaded songs:
+  - Delete (move to recycle bin)
 
 <!-- 0.3.0 -->